refactor github config compilation to seperate file
